incorporate requested changes and refactor code
